API

Index

ThreadPinning.core

ThreadPinning.cores

ThreadPinning.cpuid

ThreadPinning.cpuids

ThreadPinning.getaffinity

ThreadPinning.getcpuid

ThreadPinning.getcpuids

ThreadPinning.getnumanode

ThreadPinning.getnumanodes

ThreadPinning.hyperthreading_is_enabled

ThreadPinning.id

ThreadPinning.isefficiencycore

ThreadPinning.ishyperthread

ThreadPinning.ispinned

ThreadPinning.ncorekinds

ThreadPinning.ncores

ThreadPinning.ncputhreads

ThreadPinning.nnuma

ThreadPinning.node

ThreadPinning.nsmt

ThreadPinning.nsockets

ThreadPinning.numa

ThreadPinning.numas

ThreadPinning.openblas_getaffinity

ThreadPinning.openblas_getcpuid

ThreadPinning.openblas_getcpuids

ThreadPinning.openblas_ispinned

ThreadPinning.openblas_pinthread

ThreadPinning.openblas_pinthreads

ThreadPinning.openblas_printaffinities

ThreadPinning.openblas_printaffinity

ThreadPinning.openblas_setaffinity

ThreadPinning.openblas_setaffinity_cpuids

ThreadPinning.openblas_unpinthread

ThreadPinning.openblas_unpinthreads

ThreadPinning.pinthread

ThreadPinning.pinthreads

ThreadPinning.pinthreads_mpi

ThreadPinning.printaffinities

ThreadPinning.printaffinity

ThreadPinning.setaffinity

ThreadPinning.setaffinity_cpuids

ThreadPinning.socket

ThreadPinning.sockets

ThreadPinning.threadinfo

ThreadPinning.unpinthread

ThreadPinning.unpinthreads

ThreadPinning.visualize_affinity

ThreadPinning.with_pinthreads

References - Pinning

ThreadPinning.openblas_pinthread — Function

openblas_pinthread(cpuid; threadid)

Pin the OpenBLAS thread with the given threadid to the given CPU-thread (cpuid).

ThreadPinning.openblas_pinthreads — Function

openblas_pinthreads(cpuids; nthreads = BLAS.get_num_threads())

Pin the OpenBLAS threads to the given CPU IDs. The optional keyword argument nthreads serves as a cutoff.

ThreadPinning.openblas_setaffinity — Function

openblas_setaffinity(mask; threadid)

Set the affinity of the OpenBLAS thread with the given threadid to the given mask.

The input mask should be one of the following:

a BitArray to indicate the mask directly

a vector of cpuids (in which case the mask will be constructed automatically)

ThreadPinning.openblas_setaffinity_cpuids — Function

Set the affinity of the OpenBLAS thread to the given CPU-threads.

Examples:

openblas_setaffinity_cpuids(socket(1)) # set the affinity to the first socket

openblas_setaffinity_cpuids(numa(2)) # set the affinity to the second NUMA domain

openblas_setaffinity_cpuids(socket(1, 1:3)) # set the affinity to the first three cores in the first NUMA domain

openblas_setaffinity_cpuids([1,3,5]) # set the affinity to the CPU-threads with the IDs 1, 3, and 5.

ThreadPinning.openblas_unpinthread — Function

openblas_unpinthread(; threadid)

Unpins the OpenBLAS thread with the given threadid by setting its affinity mask to all unity. Afterwards, the OS is free to move the OpenBLAS thread from one CPU thread to another.

ThreadPinning.openblas_unpinthreads — Function

openblas_unpinthreads(; threadpool = :default)

Unpins all OpenBLAS threads by setting their affinity masks all unity. Afterwards, the OS is free to move any OpenBLAS thread from one CPU thread to another.

ThreadPinning.pinthread — Function

pinthread(cpuid::Integer; threadid = Threads.threadid())

Pin the a Julia thread to the given CPU-thread.

ThreadPinning.pinthreads — Function

pinthreads(cpuids;
    nthreads   = nothing,
    force      = true,
    warn       = is_first_pin_attempt(),
    threadpool = :default
)

Pin Julia threads to an explicit or implicit list of CPU IDs. The latter can be specified in three ways:

by passing one of several predefined symbols (e.g. pinthreads(:cores) or pinthreads(:sockets)),

by providing a logical specification via helper functions (e.g. pinthreads(numa(2, 1:4))),

explicitly (e.g. 0:3 or [0,12,4]).

See ??pinthreads for more information on these variants and keyword arguments.

Keyword arguments

If set, the keyword argument nthreads serves as a cutoff, that is, the first min(length(cpuids), nthreads) Julia threads will get pinned.

The keyword argument threadpool can be used to indicate the pool of Julia threads that should be considered. Supported values are :default (default), :interactive, or :all. On Julia >= 1.11, there is also experimental support for :gc.

If force=false, threads will only get pinned if this is the very first pin attempt (otherwise the call is a no-op). This may be particularly useful for packages that merely want to specify an (overrulable) "default pinning".

The option warn toggles general warnings, such as unwanted interference with BLAS thread settings.

Extended help

1) Predefined Symbols

:cputhreads or :compact: successively pin to all available CPU-threads.

:cores: spread threads across all available cores, only use hyperthreads if necessary.

:sockets: spread threads across sockets (round-robin), only use hyperthreads if necessary. Set compact=true to get compact pinning within each socket.

:numa: spread threads across NUMA/memory domains (round-robin), only use hyperthreads if necessary. Set compact=true to get compact pinning within each NUMA/memory domain.

:random: pin threads randomly to CPU-threads

:current: pin threads to the CPU-threads they are currently running on

:firstn: pin threads to CPU-threads in order according to there OS index.

:affinitymask: pin threads to different CPU-threads in accordance with the process affinity. By default, hyperthreads_last=true.

2) Logical Specification

The functions node, socket, numa, and core can be used to to specify CPU IDs of/within a certain domain. Moreover, the functions sockets and numas can be used to express a round-robin scatter policy between sockets or NUMA domains, respectively.

Examples (domains):

pinthreads(socket(1, 1:3)) # pin to the first 3 cores in the first socket

pinthreads(socket(1, 1:3; compact=true)) # pin to the first 3 CPU-threads in the first socket

pinthreads(numa(2, [2,4,6])) # pin to the second, the fourth, and the sixth cores in the second NUMA/memory domain

pinthreads(node(ncores():-1:1)) # pin threads to cores in reversing order (starting at the end of the node)

pinthreads(sockets()) # scatter threads between sockets, cores before hyperthreads

Different domains can be concatenated by providing them in a vector or as separate arguments to pinthreads.

Examples (concatenation):

pinthreads([socket(1, 1:3), numa(2, 4:6)])

pinthreads(socket(1, 1:3), numa(2, 4:6))

3) Explicit

Simply provide an AbstractVector{<:Integer} of CPU IDs. The latter are expected to be "physical" OS indices (e.g. from hwloc or lscpu) that start at zero!

ThreadPinning.setaffinity — Function

setaffinity(mask; threadid = Threads.threadid())

Set the affinity of a Julia thread based on the given mask (a vector of ones and zeros).

ThreadPinning.setaffinity_cpuids — Function

Set the affinity of a Julia thread to the given CPU-threads.

Examples:

setaffinity(socket(1)) # set the affinity to the first socket

setaffinity(numa(2)) # set the affinity to the second NUMA domain

setaffinity(socket(1, 1:3)) # set the affinity to the first three cores in the first NUMA domain

setaffinity([1,3,5]) # set the affinity to the CPU-threads with the IDs 1, 3, and 5.

ThreadPinning.unpinthread — Function

unpinthread(; threadid::Integer = Threads.threadid())

Unpins the given Julia thread by setting the affinity mask to all unity. Afterwards, the OS is free to move the Julia thread from one CPU thread to another.

ThreadPinning.unpinthreads — Function

unpinthreads(; threadpool::Symbol = :default)

Unpins all Julia threads by setting the affinity mask of all threads to all unity. Afterwards, the OS is free to move any Julia thread from one CPU thread to another.

ThreadPinning.with_pinthreads — Function

with_pinthreads(f::F, args...;
    soft = false,
    kwargs...
)

Runs the function f with the specified pinning and restores the previous thread affinities afterwards. Typically to be used in combination with do-syntax.

By default (soft=false), before the thread affinities are restored, the Julia threads will be pinned to the CPU-threads they were running on previously.

Example

julia> getcpuids()
4-element Vector{Int64}:
  7
 75
 63
  4

julia> with_pinthreads(:cores) do
           getcpuids()
       end
4-element Vector{Int64}:
 0
 1
 2
 3

julia> getcpuids()
4-element Vector{Int64}:
  7
 75
 63
  4

ThreadPinning.pinthreads_mpi — Function

pinthreads_mpi(symbol, rank, nranks; nthreads_per_rank, compact, kwargs...)

Pin MPI ranks, that is, their respective Julia thread(s), to (subsets of) domains (e.g. sockets or memory domains). Specifically, when calling this function on all MPI ranks, the latter will be distributed in a round-robin fashion among the specified domains such that their Julia threads are pinned to non-overlapping ranges of CPU-threads within the domain.

Valid options for symbol are :sockets and :numa.

If compact=false (default), physical cores are occupied before hyperthreads. Otherwise, CPU-cores - with potentially multiple CPU-threads - are filled up one after another (compact pinning).

The keyword argument nthreads_per_rank (default Threads.nthreads()) can be used to pin only a subset of the available Julia threads per MPI rank.

Note: As per usual for MPI, rank starts at zero.

Example:

using ThreadPinning
using MPI
comm = MPI.COMM_WORLD
nranks = MPI.Comm_size(comm)
rank = MPI.Comm_rank(comm)
pinthreads_mpi(:sockets, rank, nranks)

References - Querying

ThreadPinning.core — Function

Returns the CPU IDs that belong to core i (logical index, starts at 1). Set shuffle=true to randomize.

Optional second argument: Logical indices to select a subset of the CPU-threads.

ThreadPinning.cores — Function

Returns the CPU IDs of the system as obtained by a round-robin scattering between CPU cores. This is the same as nodes(; compact=false). Set shuffle=true to randomize.

Optional first argument: Logical indices to select a subset of the sockets.

ThreadPinning.cpuid — Function

Returns the CPU ID ("physical" OS index) that corresponds to the given logical index (starts at 1).

ThreadPinning.cpuids — Function

All valid CPU IDs of the system (sorted).

ThreadPinning.getaffinity — Function

getaffinity(; threadid = Threads.threadid(), cutoff = cpuidlimit())

Get the thread affinity of a Julia thread. Returns the affinity mask as a vector of zeros and ones. By default, the mask is cut off at Sys.CPU_THREADS. This can be tuned via the cutoff keyword argument (nothing means no cutoff).

ThreadPinning.getcpuid — Function

getcpuid(; threadid = nothing)

Returns the ID of the CPU thread on which a Julia thread is currently running.

If threadid=nothing (default), we query the id directly from the calling thread.

ThreadPinning.getcpuids — Function

getcpuids(; threadpool = :default)

Returns the IDs of the CPU-threads on which the Julia threads are currently running on.

The keyword argument threadpool (default: :default) may be used to specify a specific thread pool.

ThreadPinning.getnumanode — Function

getnumanode(; threadid = nothing)

Returns the ID (starting at zero) of the NUMA node corresponding to the CPU thread on which the calling thread is currently running. A threadid may be provided to consider a Julia thread that is different from the calling one.

ThreadPinning.getnumanodes — Function

getnumanodes(; threadpool = :default)

Returns the IDs (starting at zero) of the NUMA nodes corresponding to the CPU threads on which the Julia threads are currently running.

The keyword argument threadpool (default: :default) may be used to consider only those Julia threads that belong to a specific thread pool.

ThreadPinning.hyperthreading_is_enabled — Function

Check whether simultaneous multithreading (SMT) / "hyperthreading" is enabled.

ThreadPinning.id — Function

Returns the logical index (starts at 1) that corresponds to the given CPU ID ("physical" OS index).

ThreadPinning.isefficiencycore — Function

Returns true if the given CPU-thread lies inside of a CPU-core that has the highest power efficiency of all the CPU-cores (i.e. an efficiency value of 1). If there is only a single CPU-core kind, the return value is false for all CPU IDs.

ThreadPinning.ishyperthread — Function

Check whether the given CPU-thread is a SMT-thread / "hyperthread" (i.e. it is not the first CPU-thread in the CPU-core).

ThreadPinning.ispinned — Function

ispinned(; threadid = Threads.threadid())

Returns true if the thread is pinned, that is, if it has an affinity mask that highlights a single CPU-thread.

ThreadPinning.ncorekinds — Function

Number of different kinds of cores (e.g. efficiency and performance cores).

ThreadPinning.ncores — Function

Number of cores (i.e. excluding hyperthreads)

ThreadPinning.ncputhreads — Function

Number of CPU-threads

ThreadPinning.nnuma — Function

Number of NUMA nodes

ThreadPinning.node — Function

Returns all CPU IDs of the system/compute node (logical index, starts at 1). By default, an "cores before hyperthreads" ordering is used. Set compact=true if you want compact ordering. Set shuffle=true to randomize.

Optional second argument: Logical indices to select a subset of the CPU-threads.

ThreadPinning.nsmt — Function

The number of SMT-threads in a core. If this number varies between different cores, the maximum is returned.

ThreadPinning.nsockets — Function

Number of CPU-sockets / CPUs

ThreadPinning.numa — Function

Returns the CPU IDs that belong to the ith NUMA domain (logical index, starts at 1). By default, an "cores before hyperthreads" ordering is used. Set compact=true if you want compact ordering. Set shuffle=true to randomize.

Optional second argument: Logical indices to select a subset of the CPU-threads.

ThreadPinning.numas — Function

Returns the CPU IDs of the system as obtained by a round-robin scattering between NUMA domains. Within each NUMA domain, a round-robin ordering among CPU cores is used ("cores before hyperthreads"). Provide compact=true to get compact ordering within each NUMA domain. Set shuffle=true to randomize.

Optional first argument: Logical indices to select a subset of the sockets.

ThreadPinning.openblas_getaffinity — Function

openblas_getaffinity(; threadid, convert = true)

Query the affinity of the OpenBLAS thread with the given threadid (typically 1:BLAS.get_num_threads()). By default, returns a vector respresenting the mask. If convert=false a Ccpu_set_t is returned instead.

ThreadPinning.openblas_getcpuid — Function

openblas_getcpuid(; threadid)

Get the id of the CPU thread on which the OpenBLAS thread with the given threadid is running on according to its affinity.

Note: If the OpenBLAS thread has not been pinned before, this function will error because the affinity mask highlights more than a single CPU thread by default.

ThreadPinning.openblas_getcpuids — Function

openblas_getcpuids()

Get the ids of the CPU threads on which the OpenBLAS threads are running on according to their affinity. See openblas_getcpuid for more information."

ThreadPinning.openblas_ispinned — Function

openblas_ispinned(; threadid)

Check if the OpenBLAS thread is pinned to a single CPU thread.

ThreadPinning.openblas_printaffinities — Function

Print the affinities of all OpenBLAS threads.

ThreadPinning.openblas_printaffinity — Function

openblas_printaffinity(; threadid)

Print the affinity of an OpenBLAS thread.

ThreadPinning.printaffinities — Function

printaffinities(; threadpool = :default, kwargs...)

Print the affinity masks of all Julia threads. See printaffinity for options.

ThreadPinning.printaffinity — Function

printaffinity(; threadid::Integer = Threads.threadid())

Print the affinity mask of the Julia thread.

The keyword argument groupby may be used to change how CPU-threads are grouped visually. It defaults to groupby=:socket. Other valid values are :numa and :core.

ThreadPinning.socket — Function

Returns the CPU IDs that belong to the ith CPU/socket (logical index, starts at 1). By default, an "cores before hyperthreads" ordering is used. Set compact=true if you want compact ordering. Set shuffle=true to randomize.

Optional second argument: Logical indices to select a subset of the CPU-threads.

ThreadPinning.sockets — Function

Returns the CPU IDs of the system as obtained by a round-robin scattering between sockets. By default, within each socket, a round-robin ordering among CPU cores is used ("cores before hyperthreads"). Provide compact=true to get compact ordering within each socket. Set shuffle=true to randomize.

Optional first argument: Logical indices to select a subset of the sockets.

ThreadPinning.visualize_affinity — Function

Visualize the affinity mask of a Julia thread. Many of the keyword options of threadinfo work here as well.

ThreadPinning.threadinfo — Function

threadinfo(;
    groupby = :sockets,
    threadpool = :default,
    blas = false,
    slurm = false,
    hints = false,
    compact = true,
    hyperthreads = SysInfo.hyperthreading_is_enabled(),
    efficiency = SysInfo.ncorekinds() > 1,
    masks = false,
    coregaps = false,
    logical = false,
    color = true,
    blocksize = choose_blocksize()
)

Print information about Julia threads, e.g. on which CPU-threads (i.e. cores if hyperthreading is disabled) they are running.

Keyword arguments

groupby: Options are :sockets, :numa, :cores, or :none.

threadpool: Only consider Julia threads in the given thread pool. Supported values are :default, :interactive, and :all.

blas: Visualize BLAS threads instead of Julia threads.

slurm: Only show the part of the system that is covered by the active SLURM allocation.

hints: Give some hints about how to improve the threading related settings.

compact: Toggle between compact and "cores before hyperthreads" ordering.

hyperthreads: If true, we (try to) highlight CPU-threads that aren't the first threads within a CPU-core.

efficiency: If true, we highlight (underline) CPU-threads that belong to efficiency cores.

masks: Show the affinity masks of all Julia threads.

coregaps: Put an extra space ("gap") between different CPU-cores, when printing.

logical: Toggle between logical and "physical" CPU-thread indices.

color: Toggle between colored and black-and-white output.

blocksize: Wrap to a new line after blocksize many CPU-threads. May also be set to :numa in which case the line break will occur after each numa domain.