BLAS/LAPACK

Warning

This section isn't part of the official API. Things might change at any point without further notice.

Index

ThreadPinning.mkl_fullpath

ThreadPinning.mkl_get_dynamic

ThreadPinning.mkl_is_loaded

ThreadPinning.mkl_set_dynamic

ThreadPinning.openblas_getcpuid

ThreadPinning.openblas_getcpuids

ThreadPinning.openblas_nthreads

ThreadPinning.openblas_pinthreads

ThreadPinning.openblas_print_affinity_masks

References - OpenBLAS

ThreadPinning.openblas_getcpuid — Function

Returns the ID of the CPU thread on which the i-th OpenBLAS thread is currently running.

Note

Available as of Julia 1.9.

ThreadPinning.openblas_getcpuids — Function

Returns the IDs of the CPU-threads on which the OpenBLAS threads are currently running.

Note

Available as of Julia 1.9.

ThreadPinning.openblas_nthreads — Method

Query the number of OpenBLAS threads.

ThreadPinning.openblas_pinthreads — Method

openblas_pinthreads(
    cpuids;
    nthreads,
    juliathread,
    kwargs...
)

Pin the available OpenBLAS threads to the given CPU IDs. Currently, only explict pinning is possible.

ThreadPinning.openblas_print_affinity_masks — Function

Print the affinity masks of all OpenBLAS threads.

Note

Available as of Julia 1.9.

References - Intel MKL

ThreadPinning.mkl_fullpath — Method

mkl_fullpath(; force_update)

Returns the full path to the libmkl_rt library if the latter is loaded. Will try to locate the library and, if successfull, will cache the result. Throws an error otherwise.

To force an update of the cache, provide force_update=true.

ThreadPinning.mkl_get_dynamic — Method

mkl_get_dynamic()

Wrapper around the MKL function mkl_get_dynamic.

ThreadPinning.mkl_is_loaded — Method

Check whether Intel MKL is currently loaded via libblastrampoline (Julia >= 1.7) or is available in Libdl.dllist() (Julia 1.6).

ThreadPinning.mkl_set_dynamic — Method

mkl_set_dynamic(flag)

Wrapper around the MKL function mkl_set_dynamic.