Internals

Warning

This section isn't part of the official API. Things might change at any point without further notice.

Utility

ThreadPinning.Utility.BLAS_lib — Method

Returns the name of the loaded BLAS library (the first, if multiple are loaded).

ThreadPinning.Utility._execute — Method

Run a Cmd object, returning the stdout & stderr contents plus the exit code

ThreadPinning.Utility.affinitymask2cpuids — Method

Turn the affinity mask (a vector of ones and zeros) into a vector of CPU IDs.

ThreadPinning.Utility.cpuids2affinitymask — Method

Turn the vector of CPU IDs into an affinity mask (a vector of ones and zeros).

ThreadPinning.Utility.nblasthreads — Method

Number of BLAS threads.

ThreadPinning.Utility.taskid — Method

Returns a (most likely) unique id for the calling task.

MKL

ThreadPinning.MKL.mkl_fullpath — Method

Returns the full path to the libmkl_rt library if the latter is loaded. Will try to locate the library and, if successfull, will cache the result. Throws an error otherwise.

To force an update of the cache, provide force_update=true.

ThreadPinning.MKL.mkl_get_dynamic — Method

mkl_get_dynamic()

Wrapper around the MKL function mkl_get_dynamic.

ThreadPinning.MKL.mkl_is_loaded — Method

Check whether Intel MKL is currently loaded via libblastrampoline

ThreadPinning.MKL.mkl_set_dynamic — Method

mkl_set_dynamic(flag::Integer)

Wrapper around the MKL function mkl_set_dynamic.