Likwid-Pin

Index

ThreadPinning.likwidpin_domains

ThreadPinning.likwidpin_to_cpuids

ThreadPinning.pinthreads_likwidpin

References

ThreadPinning.likwidpin_domains — Function

likwidpin_domains(; onebased = false)

The likwid-pin compatible domains that are available for the system.

ThreadPinning.likwidpin_to_cpuids — Function

likwidpin_to_cpuids(lpstr::AbstractString; onebased = false)

Convert the given likwid-pin compatible string into a CPU ID list. See pinthreads_likwidpin for more information.

ThreadPinning.pinthreads_likwidpin — Function

pinthreads_likwidpin(str::AbstractString; onebased = false)

Pins Julia threads to CPU-threads based on the given likwid-pin compatible string. Checkout the LIKWID documentation for more information.

If the keyword argument onebased is set to true, logical indices as well as domain indices start at one instead of zero (likwid-pin default). Note, though, that this doesn't affect the explicit pinning mode where "physical" CPU IDs always start at zero.

Examples

pinthreads_likwidpin("S0:0-3")

pinthreads_likwidpin("M1:0,2,4")

pinthreads_likwidpin("S:scatter")

pinthreads_likwidpin("E:N:4:1:2")