Utility

Warning

This section isn't part of the official API. Things might change at any point without further notice.

Index

ThreadPinning.BLAS_lib

ThreadPinning.callonthreads

ThreadPinning.find_mkl

ThreadPinning.interweave

ThreadPinning.mkl_get_dynamic

ThreadPinning.mkl_is_loaded

ThreadPinning.mkl_set_dynamic

References

ThreadPinning.BLAS_lib — Method

Returns the name of the loaded BLAS library (the first, if multiple are loaded)

ThreadPinning.callonthreads — Method

Calls the function fn on all Julia threads and returns the results obtained on each thread as a ordered Vector{Any}.

Example:

julia> Pinning.callonthreads(getcpuid)
8-element Vector{Any}:
  3
  4
  5
  8
  2
 10
  6
  9

julia> pinthreads(1:Threads.nthreads())

julia> Pinning.callonthreads(getcpuid)
8-element Vector{Any}:
 1
 2
 3
 4
 5
 6
 7
 8

ThreadPinning.find_mkl — Method

Try to find libmkl_rt.so in Libdl.dllist(). Returns nothing if it can't be found.

ThreadPinning.interweave — Method

Examples

interweave([1,2,3,4], [5,6,7,8]) == [1,5,2,6,3,7,4,8]

interweave(1:4, 5:8, 9:12) == [1, 5, 9, 2, 6, 10, 3, 7, 11, 4, 8, 12]

ThreadPinning.mkl_get_dynamic — Method

mkl_get_dynamic()

Wrapper around the MKL function mkl_get_dynamic.

ThreadPinning.mkl_is_loaded — Method

Check if MKL library libmkl_rt.so is available in Libdl.dllist() (as is the case when loading MKL.jl or MKL_jll.jl).

ThreadPinning.mkl_set_dynamic — Method

mkl_set_dynamic(flag::Integer)

Wrapper around the MKL function mkl_set_dynamic.