API

Index

ThreadPinning.getcpuid

ThreadPinning.getcpuid

ThreadPinning.getcpuids

ThreadPinning.pinthread

ThreadPinning.pinthread

ThreadPinning.pinthreads

ThreadPinning.pinthreads

ThreadPinning.threadinfo

References

ThreadPinning.getcpuid — Method

Returns the ID of the CPU on which the given Julia thread (threadid) is currently executing.

ThreadPinning.getcpuid — Method

Returns the ID of the CPU on which the calling thread is currently executing.

See sched_getcpu for more information.

ThreadPinning.getcpuids — Method

Returns the ID of the CPUs on which the Julia threads are currently running.

See getcpuid for more information.

ThreadPinning.pinthread — Method

pinthread(threadid::Integer, cpuid::Integer; kwargs...)

Pin the given Julia thread (threadid) to the CPU with ID cpuid.

ThreadPinning.pinthread — Method

pinthread(cpuid::Integer; warn::Bool = true)

Pin the calling Julia thread to the CPU with id cpuid.

ThreadPinning.pinthreads — Method

pinthreads(cpuids::AbstractVector{<:Integer}[; warn])

Pins the first 1:length(cpuids) Julia threads to the CPUs with ids cpuids. Note that length(cpuids) may not be larger than Threads.nthreads().

For more information see pinthread.

ThreadPinning.pinthreads — Method

pinthreads(strategy::Symbol[; nthreads, warn, kwargs...])

Pin the first 1:nthreads Julia threads according to the given pinning strategy. Per default, nthreads == Threads.nthreads()

Allowed strategies:

:compact: pins to the first 0:nthreads-1 cores

:scatter or :spread: pins to all available sockets in an alternating / round robin fashion. To function automatically, Hwloc.jl should be loaded (i.e. using Hwloc). Otherwise, we the keyword arguments nsockets (default: 2) and hyperthreads (default: false) can be used to indicate whether hyperthreads are available on the system (i.e. whether Sys.CPU_THREADS == 2 * nphysicalcores).

:random or :rand: pins threads to random cores (ensures that no core is double occupied).

:halfcompact: pins to the first 0:2:2*nthreads-1 cores

ThreadPinning.threadinfo — Method

Print information about Julia threads, e.g. on which CPU-cores they are running.

By default, the visualization will be based on Sys.CPU_THREADS only. If you also load Hwloc.jl (via using Hwloc) it will show more detailed information.

Keyword arguments:

color (default: true): Toggle between colored and black-and-white output.

blocksize (default: 32): Wrap to a new line after blocksize many cores.

ht: If true, we highlight virtual cores associated with hyperthreads in the color=true output. By default, we try to automatically figure out whether hypterthreading is enabled.

blas (default: false): Show information about BLAS threads as well.

hints (default: false): Give some hints about how to improve the threading related settings.

ThreadPinning.@tspawnat — Macro

@tspawnat tid -> task

Mimics Base.Threads.@spawn, but assigns the task to thread tid (with sticky = true).

Example

julia> t = @tspawnat 4 Threads.threadid()
Task (runnable) @0x0000000010743c70
julia> fetch(t)
4