API

Index

ThreadPinning.cpuids_per_numa

ThreadPinning.cpuids_per_socket

ThreadPinning.getcpuid

ThreadPinning.getcpuid

ThreadPinning.getcpuids

ThreadPinning.hyperthreading_is_enabled

ThreadPinning.ishyperthread

ThreadPinning.nnuma

ThreadPinning.nsockets

ThreadPinning.pinthread

ThreadPinning.pinthread

ThreadPinning.pinthreads

ThreadPinning.pinthreads

ThreadPinning.threadinfo

References

ThreadPinning.cpuids_per_numa — Method

Returns a Vector{Vector{Int}} which indicates the CPUIDs associated with the available NUMA nodes

ThreadPinning.cpuids_per_socket — Method

Returns a Vector{Vector{Int}} which indicates the CPUIDs associated with the available CPU sockets

ThreadPinning.getcpuid — Method

Returns the ID of the CPU on which the given Julia thread (threadid) is currently executing.

ThreadPinning.getcpuid — Method

Returns the ID of the CPU on which the calling thread is currently executing.

See sched_getcpu for more information.

ThreadPinning.getcpuids — Method

Returns the ID of the CPUs on which the Julia threads are currently running.

See getcpuid for more information.

ThreadPinning.hyperthreading_is_enabled — Method

Check whether hyperthreading is enabled.

ThreadPinning.ishyperthread — Method

Check whether the given cpu thread is a hyperthread (i.e. the second cpu thread associated with a CPU-core).

ThreadPinning.nnuma — Method

Number of NUMA nodes

ThreadPinning.nsockets — Method

Number of CPU sockets

ThreadPinning.pinthread — Method

pinthread(threadid::Integer, cpuid::Integer; kwargs...)

Pin the given Julia thread (threadid) to the CPU with ID cpuid.

ThreadPinning.pinthread — Method

pinthread(cpuid::Integer; warn::Bool = true)

Pin the calling Julia thread to the CPU with id cpuid.

ThreadPinning.pinthreads — Method

pinthreads(cpuids::AbstractVector{<:Integer}[; warn])

Pins the first 1:length(cpuids) Julia threads to the CPUs with ids cpuids. Note that length(cpuids) may not be larger than Threads.nthreads().

For more information see pinthread.

ThreadPinning.pinthreads — Method

pinthreads(strategy::Symbol[; nthreads, warn, kwargs...])

Pin the first 1:nthreads Julia threads according to the given pinning strategy. Per default, nthreads == Threads.nthreads()

Allowed strategies:

:compact: pins to the first nthreads cpu threads while trying to avoid using hyperthreads (i.e. moving to next socket before using hyperthreads). If hyperthreads=true, hyperthreads will be used before moving to the next socket, if necessary.

:scatter or :spread or sockets: pins to all available sockets in an alternating / round robin fashion.

:numa: pins to all available NUMA nodes in an alternating / round robin fashion.

:random or :rand: pins threads to random cpu threads (ensures that no cpu thread is double occupied). By default (hyperthreads=false), hyperthreads will be ignored.

:firstn: pins to the cpuids 0:nthreads-1

ThreadPinning.threadinfo — Method

Print information about Julia threads, e.g. on which cpu threads (i.e. cores if hyperthreading is disabled) they are running.

Keyword arguments:

color (default: true): Toggle between colored and black-and-white output.

blocksize (default: 32): Wrap to a new line after blocksize many cpu threads.

hyperthreading (default: true): If true, we (try to) highlight cpu threads associated with hyperthreading in the color=true output.

blas (default: false): Show information about BLAS threads as well.

hints (default: false): Give some hints about how to improve the threading related settings.

groupby (default: :sockets): Options are :sockets, :numa, or :none.