API

Index

ThreadPinning.cpuids_all

ThreadPinning.cpuids_per_core

ThreadPinning.cpuids_per_node

ThreadPinning.cpuids_per_numa

ThreadPinning.cpuids_per_socket

ThreadPinning.getcpuid

ThreadPinning.getcpuid

ThreadPinning.getcpuids

ThreadPinning.hyperthreading_is_enabled

ThreadPinning.ishyperthread

ThreadPinning.ncores

ThreadPinning.ncores_per_numa

ThreadPinning.ncores_per_socket

ThreadPinning.ncputhreads

ThreadPinning.ncputhreads_per_core

ThreadPinning.ncputhreads_per_numa

ThreadPinning.ncputhreads_per_socket

ThreadPinning.nnuma

ThreadPinning.nsockets

ThreadPinning.pinthread

ThreadPinning.pinthread

ThreadPinning.pinthreads

ThreadPinning.pinthreads

ThreadPinning.print_affinity_masks

ThreadPinning.sysinfo

ThreadPinning.threadinfo

ThreadPinning.unpinthread

ThreadPinning.unpinthreads

References

ThreadPinning.cpuids_all — Method

Returns a Vector{Int} which lists all valid CPUIDs. There is no guarantee about the order except that it is the same as in lscpu.

ThreadPinning.cpuids_per_core — Method

Returns a Vector{Vector{Int}} which indicates the CPUIDs associated with the available physical cores

ThreadPinning.cpuids_per_node — Method

Returns a Vector{Int} which indicates the CPUIDs associated with the available node. Physical cores come first. Set compact=true to get compact ordering.

ThreadPinning.cpuids_per_numa — Method

Returns a Vector{Vector{Int}} which indicates the CPUIDs associated with the available NUMA domains. Within each memory domain, physical cores come first. Set compact=true to get compact ordering instead.

ThreadPinning.cpuids_per_socket — Method

Returns a Vector{Vector{Int}} which indicates the CPUIDs associated with the available CPU sockets. Within each socket, physical cores come first. Set compact=true to get compact ordering instead.

ThreadPinning.getcpuid — Method

Returns the ID of the CPU on which the given Julia thread (threadid) is currently executing.

ThreadPinning.getcpuid — Method

Returns the ID of the CPU on which the calling thread is currently executing.

See sched_getcpu for more information.

ThreadPinning.getcpuids — Method

Returns the ID of the CPUs on which the Julia threads are currently running.

See getcpuid for more information.

ThreadPinning.hyperthreading_is_enabled — Method

Check whether hyperthreading is enabled.

ThreadPinning.ishyperthread — Method

Check whether the given cpu thread is a hyperthread (i.e. the second cpu thread associated with a CPU-core).

ThreadPinning.ncores — Method

Number of cores (i.e. excluding hyperthreads)

ThreadPinning.ncores_per_numa — Method

Number of CPU cores per NUMA domain

ThreadPinning.ncores_per_socket — Method

Number of CPU cores per socket

ThreadPinning.ncputhreads — Method

Number of CPU threads

ThreadPinning.ncputhreads_per_core — Method

Number of CPU threads per core

ThreadPinning.ncputhreads_per_numa — Method

Number of CPU threads per NUMA domain

ThreadPinning.ncputhreads_per_socket — Method

Number of CPU threads per socket

ThreadPinning.nnuma — Method

Number of NUMA nodes

ThreadPinning.nsockets — Method

Number of CPU sockets

ThreadPinning.print_affinity_masks — Method

Print the affinity masks of all Julia threads.

ThreadPinning.sysinfo — Method

Get information about the system like how many sockets or NUMA nodes it has, whether hyperthreading is enabled, etc.

ThreadPinning.pinthread — Method

pinthread(threadid::Integer, cpuid::Integer; kwargs...)

Pin the given Julia thread (threadid) to the CPU with ID cpuid.

ThreadPinning.pinthread — Method

pinthread(cpuid::Integer; warn::Bool = true)

Pin the calling Julia thread to the CPU with id cpuid.

ThreadPinning.pinthreads — Method

pinthreads(cpuids::AbstractVector{<:Integer}[; warn])

Pins the first 1:length(cpuids) Julia threads to the CPUs with ids cpuids. Note that length(cpuids) may not be larger than Threads.nthreads().

For more information see pinthread.

ThreadPinning.pinthreads — Method

pinthreads(pinning[; places, nthreads=Threads.nthreads(), warn=true, force=true, kwargs...])

Pins the first 1:nthreads Julia threads according to the given pinning strategy to the given places. Per default, a reasonable value for places is chosen based on the given pinning strategy.

Pinning strategies (pinning):

:compact or :close: pins to places one after another.

:spread or scatter: pins to places in an alternating / round robin fashion.

:random: shuffles the given places and then pins to them compactly.

:current: pins threads to the cpu threads where they are currently running (ignores places).

:firstn: pin threads to the first nthreads() cpu ids reported by lscpu.

Places (places):

:cores or Cores(): all the cores of the system

:threads, :cputhreads, or CPUThreads(): all the cpu threads of the system (equal to :cores if there is one cpu thread per core, e.g. no hyperthreading)

:sockets or Sockets(): the sockets of the system

:numa or NUMA(): the NUMA domains of the system

An AbstractVector{<:AbstractVector{<:Integer}} of cpu ids that defines the places explicitly

If force=false the pinthreads call will only pin threads if this is the first attempt to pin threads with ThreadPinning.jl. Otherwise it will be a no-op. This may be particularly useful for packages that merely want to specify a "default pinning".

ThreadPinning.unpinthread — Method

Unpins the Julia thread with the given threadid by setting the affinity mask to all unity. Afterwards, the OS is free to move the Julia thread from one CPU thread to another.

ThreadPinning.unpinthreads — Method

Unpins all Julia threads by setting the affinity mask of all threads to all unity. Afterwards, the OS is free to move any Julia thread from one CPU thread to another.

ThreadPinning.threadinfo — Method

Print information about Julia threads, e.g. on which cpu threads (i.e. cores if hyperthreading is disabled) they are running.

Keyword arguments:

color (default: true): Toggle between colored and black-and-white output.

blocksize (default: 32): Wrap to a new line after blocksize many cpu threads.

hyperthreading (default: true): If true, we (try to) highlight cpu threads associated with hyperthreading in the color=true output.

blas (default: false): Show information about BLAS threads as well.

hints (default: false): Give some hints about how to improve the threading related settings.

groupby (default: :sockets): Options are :sockets, :numa, :cores, or :none.

masks (default: false): Show the affinity masks of all Julia threads.