Utility

Warning

This section isn't part of the official API. Things might change at any point without further notice.

Index

ThreadPinning.BLAS_lib

ThreadPinning._openblas_setaffinity

ThreadPinning.find_mkl

ThreadPinning.interweave

ThreadPinning.mkl_get_dynamic

ThreadPinning.mkl_is_loaded

ThreadPinning.mkl_set_dynamic

ThreadPinning.openblas_set_affinity_mask

References

ThreadPinning.interweave — Method

Examples

interweave([1,2,3,4], [5,6,7,8]) == [1,5,2,6,3,7,4,8]

interweave(1:4, 5:8, 9:12) == [1, 5, 9, 2, 6, 10, 3, 7, 11, 4, 8, 12]

ThreadPinning.@tspawnat — Macro

@tspawnat tid -> task

Mimics Base.Threads.@spawn, but assigns the task to thread tid (with sticky = true).

Example

julia> t = @tspawnat 4 Threads.threadid()
Task (runnable) @0x0000000010743c70
julia> fetch(t)
4

ThreadPinning.BLAS_lib — Method

Returns the name of the loaded BLAS library (the first, if multiple are loaded)

ThreadPinning._openblas_setaffinity — Method

Sets thread affinity for OpenBLAS threads. thread_idx is in [0, openblasgetnum_threads()-1].

ThreadPinning.find_mkl — Method

Try to find libmkl_rt.so in Libdl.dllist(). Returns nothing if it can't be found.

ThreadPinning.mkl_get_dynamic — Method

mkl_get_dynamic()

Wrapper around the MKL function mkl_get_dynamic.

ThreadPinning.mkl_is_loaded — Method

Check if MKL library libmkl_rt.so is available in Libdl.dllist() (as is the case when loading MKL.jl or MKL_jll.jl).

ThreadPinning.mkl_set_dynamic — Method

mkl_set_dynamic(flag::Integer)

Wrapper around the MKL function mkl_set_dynamic.

ThreadPinning.openblas_set_affinity_mask — Method

The input mask should be either of the following:

a BitArray indicating the mask directly

a vector of cpuids (the mask will be constructed automatically)